Make the GoogleSearch provider understand action.WEB_SEARCH
The search dialog & browser are going to use the system default handler for action.WEB_SEARCH and use that for suggestions & launching web searches. This change makes GoogleSearch provider compatible with that model so it can be used for google searches and suggestions. With this change, Google Search also appears as an item in the Web Search providers list in system search settings.