Use ACTION_SEARCH to launch search URLs.
This was done so that the browser can identify them as search requests and add location information to the request if enabled.

Also adding internet permission to the manifest as without that it was throwing errors in the debug log infrequently (we need it anyway since this provider accesses the internet for google searches, but in the past it was running inside the acore process so perhaps was not required and this design may have changed now).